[refactor] 전체 파일 구조 변경 및 router 추가
