+++ src/pages/Card/index.ts
@@ -1,0 +1,5 @@
+import CardCompletePage from './CardCompletePage';
+import CardIndexPage from './CardIndexPage';
+import CardNewPage from './CardNewPage';
+
+export { CardIndexPage, CardNewPage, CardCompletePage };
