+++ src/utils/cardUtil.ts
