+++ tsconfig.json
@@ -22,11 +22,12 @@
     "noEmit": true,
     "jsx": "react-jsx",
     "paths": {
-      "@constants": ["src/common/constants/index"],
+      "@constants": ["src/constants/index"],
       "@components": ["src/components/index"],
-      "@interface": ["src/common/interface/index"],
+      "@interface": ["src/interface/index"],
       "@share": ["src/components/share/index"],
-      "@utils": ["src/common/utils/index"],
+      "@utils": ["src/utils/index"],
+      "@hooks": ["src/hooks/index"],
     }
   },
   "include": [
